feat: combine layout context and provider
